add useNavigation.ts, move AppNavigation to InstancePageTemplate.tsx
